⚡️feat: [DEMO-135] #comment Cost V2 적용 - 24.03.30
#close #time 1m +review SR @dumbbelloper

See https://dumbbelloper.atlassian.net/browse/DEMO-135 more details.

[ Frontend-Web::Java-Service-Tree-Framework(JSTF)::313DEVGRP ]
요구 관리 : [Requirement Manage] http://www.a-rms.net
문서 관리 : [Document] http://www.313.co.kr/confluence
이슈 관리 : [IssueTracker] http://www.313.co.kr/jira
소스 관리 : [VersionControl] https://github.com/313devgrp
코드 리뷰 : [CodeReview] http://www.313.co.kr/fecru
빌드 관리 : [CICD-Deploy] http://www.313.co.kr/jenkins
배포 관리 : [Deploy-Pipeline] http://www.313.co.kr/spinnaker
아티 팩트 : [ArtifactManager] http://www.313.co.kr/nexus
코드 분석 : [CodeAnalysis] http://www.313.co.kr/sonar
성능 측정 : [ZipKin] http://www.313.co.kr/zipkin/

도커 관리 : [Docker Hub] https://hub.docker.com/u/313devgrp
클러스터1 : [Kubernetes] http://www.313.co.kr:31380
클러스터2 : [DockerSwarm] http://www.313.co.kr/portainer/#!/auth

디비 관리 : [DB Admin] http://www.313.co.kr/phpmyadmin/
오브 젝트 : [S3 Admin] http://www.313.co.kr/minio/login
캐시 관리 : [Cache Admin] http://www.313.co.kr/redis/
ELFK 스택1 : [NoSql Index] http://www.313.co.kr/elasticsearch/_cat/indices
ELFK 스택2 : [NoSql Node] http://www.313.co.kr/elasticsearch/_nodes?pretty=true
ELFK 스택3 : [LogStash] http://www.313.co.kr/logstash/_node/stats?pretty
ELFK 스택4 : [Kibana] http://www.313.co.kr/kibana/app/home#/

모니터링 1 : [ElasticHQ] http://www.313.co.kr/elastichq/#!/
모니터링 2 : [Grafana] http://www.313.co.kr/grafana

인증 인가 : [Auth] http://www.313.co.kr/auth/

NFS 관리 : [NAS] http://www.313.co.kr:5000/webman/index.cgi
메일 관리 : [Mail] http://www.313.co.kr/mail/
원격 관리 : [RDP] http://www.313.co.kr/guacamole/#/
