--- arms/html/analysisCostV2/content-header.html
@@ -1,14 +1,0 @@
-<h3 class="page-title font16">
-	<i class="fa fa-bar-chart-o"></i> Analysis
-	<small>CostV2</small>
-	<ol class="breadcrumb pull-right hidden-xs font12">
-		<li>
-			<a href="#">
-				<i class="fa fa-sitemap"></i>
-				Home
-			</a>
-		</li>
-		<li class="active">Analysis</li>
-		<li class="active">CostV2</li>
-	</ol>
-</h3>
